[AIRFLOW-6344] Fix travis CI for tag builds
Don't try to find changed files unless we are building a pull request.
This only caused a problem on build of tags, but we were also doing this
for master/branch builds, but it was always saying finding no files
changed.

By checking this early we can make the other conditions in this function
simpler.

(cherry picked from commit 67463c3d8e5fe1618117244364d8a49f80536820)