Bugfix: Check next run exists before reading data interval (#19307)
Fix #19304, and also an issue on scheduling a DAG's first-ever run introduced in #18897. We could fix it outside this function, but if `next_dagrun` is None, the next run's data interval is supposed to be None in the first place, so checking inside this function just makes sense.

closes https://github.com/apache/airflow/issues/19343
closes https://github.com/apache/airflow/issues/19304

(cherry picked from commit dc4dcaa9ccbec6a1b1ce84d5ee42322ce1fbb081)