[astro] Fix graph view states (#17821)
When https://github.com/apache/airflow/pull/17821 was cherry-picked into
OSS 2.1.4, the `for` (v2-1-stable) vs `forEach` (main) wasn't accounted
for, and this changed the behavior of the method.

https://github.com/apache/airflow/pull/18607 attempted to fix it, but
ultimately didn't change anything because main wasn't actually broken.

The solution is to `continue` so we finish the loop, not short-circuit
out with `return`.